add manage boook and people
